register inside submodules
